% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipecomposer.R
\name{import_clipboard}
\alias{import_clipboard}
\title{Use Pipe Composer on code pipeline from system clipboard}
\usage{
import_clipboard(...)
}
\arguments{
\item{...}{Parameters to pass through to \code{\link{pipecomposer}}.}
}
\description{
Uses the system clipboard to prepopulate and launch the pipe composer
interface. The clipboard must contain a single, valid, pipe-separated
expression.
}

