% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipecomposer.R
\name{pipecomposer}
\alias{pipecomposer}
\title{Visualize pipelines of dplyr/ggvis commands}
\usage{
pipecomposer(stage_count = getOption("pipecomposer.stage_count", 8),
  indent = getOption("pipecomposer.indent", "  "))
}
\arguments{
\item{stage_count}{Number of stages to display.}

\item{indent}{String to use to indent stages two and later when printing code
  to the console.}
}
\description{
Launches a browser-based user interface for composing chains of piped
commands, especially ones that return data frames or \code{ggvis} objects.
For more about pipes, see \link[magrittr]{magrittr}.
}

